backend/requirement.txt
Django==2.2.1
djangorestframework==3.9.4
psycopg2==2.8.3
dj_database_url==0.3.0
africastalking==1.1.6
phonenumbers==8.9.14
python-dateutil==2.7.3
xlrd==1.2.0
openpyxl==2.6.2
celery==4.4.5
redis==3.5.3
